Prefer lower ids on ties
This will change in the next ELKI release. It was intended to prefer lower IDs, and this "bug" (it's arbitrary, so it is not really a bug) has been fixed now.
Lower ids yields more expected results. In particular, if you reorder the data set to be in cluster order, and rerun OPTICS, the result should now be stable.